move table method from api_database.go to api_table.go
